PyDocFlow-Studio/requirements.txt
# Core Streamlit framework
streamlit>=1.28.0

# PDF processing libraries
PyMuPDF>=1.23.0
pdfminer.six>=20221105
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# EPUB processing libraries
ebooklib>=0.18
python-docx>=1.0.0
reportlab>=4.0.0
lxml>=4.9.3

# Web scraping and data processing
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0

# Additional utilities
python-dateutil>=2.8.2